Demo change
I gave an example of a better way to write that html. By breaking it up, we can make more sense of it by looking at the html. This will also help any search engine understand that your <h4> means something (a job title) instead of being spammed on the page.